requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==23.0.0

pandas==2.2.1
numpy==1.26.4

pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==12.1.1

openpyxl==3.1.2
xlrd==2.0.1

python-multipart==0.0.22
python-dotenv==1.0.1

reportlab==4.1.0

sqlalchemy==2.0.29
psycopg2-binary==2.9.9

python-docx==1.1.0
